Tests for Loader

+++ src/main/java/com/spikeify/aerospike/ByteConverter.java
@@ -4,18 +4,18 @@ import com.spikeify.Converter;
 
 import java.lang.reflect.Type;
 
-public class ByteConverter implements Converter<Byte, Integer> {
+public class ByteConverter implements Converter<Byte, Long> {
 
 	public boolean canConvert(Class type) {
 		return Byte.class.isAssignableFrom(type) || byte.class.isAssignableFrom(type);
 	}
 
-	public Byte fromProperty(Integer property) {
+	public Byte fromProperty(Long property) {
 		return property.byteValue();
 	}
 
-	public Integer fromField(Byte fieldValue) {
-		return Integer.valueOf(fieldValue);
+	public Long fromField(Byte fieldValue) {
+		return Long.valueOf(fieldValue);
 	}
 
 }
